feat: Lab room implemented, multiple features
This commit adds multiple features. The lab room has a potion recipe that is randomised each time. The cauldron can be clicked so that ingredients can be dragged into the cauldron to 'brew' a potion. The lever in the middle of the room must be clicked and then the jewellery box to get to the riddle. After the riddle is solved the user 'collects' the crystal. Note, I added sounds but not to this commit because im unsure if we can commit pom files, The blurred images will be unblurred when other objects are found, the images I used for the ingredients are just placeholders.