feat: Implemented matching tile game
This commit adds a matching tile game which can be played in the dragon room. The user needs to play and win this mini game in order to access the object in the dragon room. Related images have been added. A tile class has been created for the tile game. A controller and fxml file has been created for the mathcing game. The dragon room has been made accessible by clicking the initial room window for now.