fix: Tiles wait some time before dissapearing
This commit fixes the matching game. Now when the user does not match a pair of tiles correctly, it waits a certain amount of time before dissapearing so that the user has time to see the tiles selected. Also, the user cannot click other tiles during this time. Styling changes have also been made, most notably the button.