feat: Implemented game win
This commit adds the functionality of winning the game. Once the user has solved the riddle, pulled the lever, clicked the jewel box and collected the ingredient in the lab, along with other room ingredients, the user can then combine the ingredients in the couldron. Once the user has combined the ingredients correctly, they can click on the window to escape. This brings up the win screen.