fix: fixed ai for easy and hard mode
This commit fixes bugs for easy and hard mode of the ai chat. In easy mode messages were previously appended if the riddle had been solved and the user asked for a hint. This has been fixed. In hard mode, previously the ai would not tell the user if their answer to the riddle is wrong. This has been fixed